Move documentation into the repository
Currently there is a problem in which users would like to contribute
changes to the documentation. The reasons for this are multitude, but
more generally:

- Add additional helpful information to the wiki (#1926, #1756, #1165)
- Add additional documentation describing the purpose of sniffs (#1603)

This commit moves the documentation into the source repository in the
"docs" folder, as well as adds configuration for the documentation tool
"mkdocs"

== Design Notes ==

=== Why Mkdocs ===

There are many different documentation and static sites generation
tools. However, mkdocs has some uniquely nice properties:

  - It's extremely simple
  - It defines strict conventions (all content in docs folder)
  - It allows generating HTML documentation sites from otherwise
    well structured markdown documentation
  - It's in fairly wide use[1]

=== Move "generic" documentation into subfolder ===

There are broadly two classes of documentation that are anticipated to
be part of this repository:

  - "How to" documentation
  - "Reference" documentation

The former is the "generic" documentation, and includes things like
usage, configuration and so fourth.

The latter is per-sniff documentation describing the purpose of each
sniff.

The existing directory structure factors in both types. Future work
may seek to reorganise the documentation further, but this gets the
pattern started with minimal overhead.

[1] - https://github.com/search?q=mkdocs.yml