Add an example Sniff
This commit introduces documentation for an example sniff; one that is
not otherwise documented within the repository. It includes examples of
what is good, what is not good and an explanation of what the sniff is
supposed to resolve.

This is not designed to be an absolute guide on what is supposed to be
within a sniff, but rather a guide on what pattern the sniffs should
take.

By naming the sniff documentation after the reference produced with:

  ./phpcs.phar --standard=PSR2  --report="json"

It's possible to link directly to this documentation within PHPCS,
or at least allow users to "guess" the URL correctly.

== Design Notes ==

=== Not using generated markdown ===

A far better pattern for this would be to continue the pattern of
documenting this within PHPCS itself, and generating the appropriate
documentation in the build step.

However, such documentation appears incomplete, and is generated in a
way that does not make it easily suitable for the aforementioned
documentation structure.

Future work might seek to implement a new "Mkdocs" generator that
generates documentation for all sniffs, to be executed in the build
step.

=== Delimited by `/` and not `.` ===

Initially the URL was to be formed like:

  sniffs/psr2.classes.classdeclaration.openbracenewline

However, this made for a large pool of sniffs that were not inherently
related all together in one large menu, or on disk.

The revised solution is to delimit the sniffs where the sniffs are
already delimited via the `.` character as separate folders.

This allows URLs like:

  sniffs/psr2/classes/classdeclaration/openbracenewline/

Which are still trivial to generate, but make for a much better
organisation both in menus and on disk.