Test bootstrap/printPHPCodeSnifferTestOutput(): bug fix
PR 2289 added a count of the number of test case files to the test output, however did not take into account that when only the unit tests for the utility functions would be run, i.e. `phpunit --filter FindExtendedClassNameTest`, the `$GLOBALS['PHP_CODESNIFFER_SNIFF_CASE_FILES']` variable would not have any arrays in it.

This adds some minor defensive coding which gets rid of the error messages which would be generated in that situation.