Refactor: add new Util\Sniffs\Conditions class
Move the methods related to condition checking out of the `File` class to a separate class with static methods.

This deprecates the `File::hasCondition()` and the `File::getCondition()` methods in favour of the same in a new `Util\Sniffs\Conditions` class.

This functionality is only used by a small percentage of sniffs, so is better placed in a dedicated class.
This also allows for additional condition related functions to be added to this dedicated class.

The deprecated methods should be removed in PHPCS 4.0.