Generic.CodeAnalysis.UnusedFunctionParameter: improve accuracy / bug fix
This commit is two-fold:

On the one hand, it improves handling of anonymous classes which extend another class or implement an interface.
This means that the error code for those errors will now correctly indicate that the _unused function parameter_ was found in an extended class or implemented interface.

On the other hand, the sniff now recognizes that classes can be nested and will only look at the actual class in which the method is found, not at a higher level class.
This fixes a bug were the error code would be incorrect for unused parameters in nested classes.

Includes unit tests for both cases.