Refactor/Conditions: implement use of the new utility methods in File::getMemberProperties()
... without changing the existing behaviour.

It could be argued that the parse error warning can now be removed in favour of just throwing the Exception, but that would be a BC-break for sniffs using this method and expecting to receive an empty array instead of the Exception.

I'd suggest making that change in PHPCS 4.

Alternatively, the parse error warning could already be removed with just an empty array being returned for now.