Zend/ValidVariableName: fix faulty unit test
Interfaces cannot declared properties. Traits however can and were so far not tested.

The switch over to the `Conditions::isOOProperty()` method in the `AbstractVariableSniff` means that a variable declared as a property in an interface will no longer be seen as a property and therefore the warning which would be generated by the `File::getMemberProperties()` method will no longer be thrown.

However, if the construct is changed to a `trait`, the normal logic will kick in and the `PrivateNoUnderscore` error will be thrown.