Refactor: add new Util\Sniffs\ObjectDeclarations class
Move the methods related to the checking of object declaration statements out of the `File` class to a separate class with static methods.

This deprecates the `File::getClassProperties()`, `File::findExtendedClassName()` and the `File::findImplementedInterfaceNames()` methods in favour of the same in a new `Util\Sniffs\ObjectDeclarations` class.

This functionality is only used by a small percentage of sniffs, so is better placed in a dedicated class.
This also allows for additional object declaration related functions to be added to this dedicated class.

The deprecated methods should be removed in PHPCS 4.0.

Existing unit tests for these methods have been moved to the `tests\Core\Util\Sniffs\ObjectDeclarations` directory.

Includes minor documentation fix in the test function docblocks.
Includes renaming a test method parameter for consistency with other utility method unit tests.