Refactor/ObjectDeclarations: refactor findExtendedClassName/findExtendedInterfaceNames methods
As the logic needed for these two methods, as well as for the (upcoming) `findImplementedInterfaces()` method, is basically the same, I've extracted the logic out to a private `ObjectDeclarations::findExtendedImplemented()` method which is now called by both methods to DRY out the code (remove code duplication).

This refactor includes a functional change.

Previously, the methods would use the `File::getTokensAsString()` method to retrieve the name of extended/implemented objects.
The name, however, may contain superfluous whitespace and/or comments which would invalidate any comparison against the name.

To that end, the name is now returned "clean", i.e. stripped of all whitespace and/or comments.

Includes additional unit tests for the changed functionality, as well as for some edge cases.