Refactor: add new Util\Sniffs\FunctionDeclarations class
Move the methods related to the checking of function declaration statements out of the `File` class to a separate class with static methods.

This deprecates the `File::getMethodParameters()` and the `File::getMethodProperties()` methods in favour of the `getParameter()` and `getProperties()` methods in a new `Util\Sniffs\FunctionDeclarations` class.

This functionality is only used by a small percentage of sniffs, so is better placed in a dedicated class.
This also allows for additional function declaration related functions to be added to this dedicated class.

Take note of the dropping of `Method` from the method names. As these methods also work for global functions as well as closures and these methods are being moved (renamed) now anyway, it seemed appropriate to make the names more apt.

The deprecated methods should be removed in PHPCS 4.0.

Existing unit tests for these methods have been moved to the `tests\Core\Util\Sniffs\FunctionDeclarations` directory.

Includes renaming a test method parameter for consistency with other utility method unit tests.