Refactor/FunctionDeclarations/GetParameters: add additional unit tests
Add various additional unit tests to make sure the method is completely tested for all cases the method accounts for.

Includes adding some additional unit tests documenting the behaviour of the function with fully qualified class names as type declaration, including when such a type declaration is interlaced with comments and whitespace.