Refactor/FunctionDeclarations: implement in Squiz.Functions.GlobalFunctions / bug fix
Implement the use of the new `FunctionDeclarations::isMagicFunction()` method in the `Squiz.Functions.GlobalFunctions` sniff.

This includes re-ordering and improving the logic in the sniff by using `Conditions::hasCondition()` instead of checking that the token conditions are empty.
This fixes a bug where a conditional function declaration in the global namespace would not be recognized as a global function declaration.

The `if ($functionName === null)` can be considered redundant as the sniff doesn't look at closures, so the chances of that condition ever being true are extremely slim. I've left it in just in case, as it doesn't do any harm either.

* Includes unit test for the bug fix.
* Includes fixing an unintentional parse error in the unit test case file.