Refactor: add new Util\Sniffs\Variables class
Move the methods related to the checking of variables out of the `File` class to a separate class with static methods.

This deprecates the `File::getMemberProperties()` method in favour of the same in the new `Util\Sniffs\Variables` class.

This functionality is only used by a small percentage of sniffs, so is better placed in a dedicated class.
This also allows for additional variable declaration related functions to be added to this dedicated class.

The deprecated features should be removed in PHPCS 4.0.

Existing unit tests for these methods have been moved to the `tests\Core\Utils\Sniffs\Variables` directory.

Includes minor documentation fix in test function docblock.
Includes renaming a test method parameter for consistency with other utility method unit tests.