Refactor/UseStatements: implement isImportUse() in PSR2.Namespaces.UseDeclaration / bug fix
Implement the new `UseStatements::isImportUse()` method in the `PSR2.Namespaces.UseDeclaration` sniff.

This should improve the consistency and accuracy of the results and fixes a bug where a trait use statement in an anonymous class would inadvertently be seen as an import use statement, resulting in a false positive.

Includes unit test for the bug.