Refactor/Namespaces: implement getDeclaredName() in Generic/DuplicateClassName / various bug fixes
Implement the use of the new `Namespaces::getDeclaredName()` method and the `Namespaces::$statementClosers` property in the `Generic.Classes.DuplicateClassName` sniff.

This should improve the consistency and accuracy of the results and fixes a number of different bugs.
1. It fixes a nasty bug which could throw the sniff into an eternal loop when an unfinished namespace statement was encountered during live coding.
2. It also fixes false positives for when a namespace _operator_ was encountered.
3. It fixes false negatives for when a namespace declaration interlaced with comments was encountered.

Includes unit tests for all bugs.

Additionally, a unit test for classes declared within a scoped global namespace has also been added.

**Note**: the unit test for the live coding bug has to be the last unit test file for this sniff so as to not influence the other tests. To that end, it has been numbered `99`.