Refactor/Namespaces: implement determineNamespace() in PSR1/ClassDeclaration / various bug fixes
Implement the use of the new `Namespaces::determineNamespace()` method in the `PSR1.Classes.ClassDeclaration` sniff.

This should improve the consistency and accuracy of the results and fixes a number of bugs.

Bugs fixed:
1. A false negative for a class declared in a non-namespaced file with a function call using the namespace operator before the class declaration.
2. False negative for a class declared in an explicit global namespace.
3. False negative for a class declared in an explicit global namespace while there are multiple namespace declarations in the files.

Includes unit test for the bugs.

The fixed bugs should also improve the accuracy of the metrics this sniff records.