Refactor/Namespaces: implement isDeclaration() in PSR1/SideEffects / bug fix
Implement the use of the new `Namespaces::isDeclaration()` method in the `PSR1.Files.SideEffects` sniff.

This should improve the consistency and accuracy of the results and fix a bug where there would be a false negative for a class declared before or after a function call using the namespace operator.

Includes unit tests for the bug.