Refactor/Namespaces: implement isDeclaration() in PSR2/NamespaceDeclaration / bug fix
Implement the use of the new `Namespaces::isDeclaration()` method in the `PSR2.Namespaces.NamespaceDeclaration` sniff.

This should improve the consistency and accuracy of the results and fix a bug where there would be a false positive for a function call using the namespace operator.

Includes unit tests for the bug.