Refactor/Namespaces: implement in PSR2/UseDeclaration / bug fix
Implement the use of the new `Namespaces::isDeclaration()` and `Namespaces::findNamespacePtr()` methods in the `PSR2.Namespaces.UseDeclaration` sniff.

This should improve the consistency and accuracy of the results and fixes a few bugs where the sniff would get confused over the namespace keyword being used as an operator.

Includes unit tests for the bugs.