Refactor/Namespaces: implement determineNamespace() in Squiz/SelfMemberReference / bug fix
Implement the use of the new `Namespaces::determineNamespace()` method in the `Squiz.Classes.SelfMemberReference` sniff.

This should improve the consistency and accuracy of the results and fixes a false negative where the sniff would get confused over the namespace keyword being used as an operator between the actual namespace declaration and the self member reference, as well as a false negative when the namespace declaration ends with a PHP close tag.

Includes unit tests for the bugs, as well as a unit test verifying the correct behaviour of the sniff when in an explicit global namespace.