Refactor: add new Util\Sniffs\TokenIs class
Move methods related to dealing with ambiguous tokens out of the `File` class to a separate class with static methods.

This deprecates the `File::isReference()` method in favour of the same in a new `Util\Sniffs\TokenIs` class.

This functionality is only used by a small percentage of sniffs, so is better placed in a dedicated class.
This also allows for additional functions dealing with ambiguous tokens to be added to this dedicated class.

The deprecated features should be removed in PHPCS 4.0.

The associated unit tests for the `isReference()` method have been moved to the `tests/Core/Util/Sniffs/TokenIs` directory.
Includes renaming a test method parameter for consistency with other utility method unit tests.