Refactor/TokenIs: implement isUnaryPlusMinus() in Squiz/OperatorSpacing / bug fix
Implement the use of the new `TokenIs::isUnaryPlusMinus()` method in the `Squiz.WhiteSpace.OperatorSpacing` sniff.

This should improve the consistency and accuracy of the results and fix a bug where spacing would be added if a unary operator was preceded by a cast token.

Includes unit test for the bug.