Refactor/TokenIs: implement isUnaryPlusMinus() in Squiz/OperatorBracket / enhancement
Implement the use of the new `TokenIs::isUnaryPlusMinus()` method in the `Squiz.Formatting.OperatorBracket` sniff.

This should improve the consistency and accuracy of the results.

It also fixes an oversight where a space between a unary minus sign and a number _would_ be reported, but a space between a unary plus sign and a number would **not** be reported.

As the error message and code for this issue were very "minus" specific, it has been adjusted to be slightly more generic.

To make this also work for JS files, the check for the "`T_PLUS` confusion prevention between operator and concatenation" has been moved to _after_ the check for the unary sign.

**Note**: This does constitute a minor BC-break as the errorcode changes from `SpacingAfterMinus` to `SpacingAfterSign`.

Includes unit test for the fix.