Refactor: add new Util\Sniffs\ConstructNames class
Move the methods related to checking the names of structures out of the `File` and the `Common` class to a separate class with static methods.

This deprecates the `File::getDeclarationName()`, the `Common::isCamelCaps()` and the `Common::isUnderscoreName()` methods in favour of the same in a new `Util\Sniffs\ConstructNames` class.

This functionality is only used by a small percentage of sniffs, so is better placed in a dedicated class.
This also allows for additional construct name related functions to be added to this dedicated class.

The deprecated features should be removed in PHPCS 4.0.

Associated unit tests for the `isCamelCaps()` method have been appropriately adjusted.

Includes minor docs fix for the `@throws` tag documentation for the `getDeclarationName()` method.