Refactor/ConstructNames: various improvements to the getDeclarationName() method
While certain names may be invalid, such as construct names starting with a number; for the purpose of alerting users to these invalid names, the `getDeclarationName()` method should be able to pick up on them.

The `getDeclarationName()` method has now been adjusted to allow for such names.

The way this has been implemented also improves the parse error resilience for the `getDeclarationName()` method.

The `getDeclarationName()` method could previously inadvertently misidentify the name of a structure in the case of parse errors/live coding as it looked for the first `T_STRING` after the keyword without limiting it to the declaration statement. This has now also been fixed.

To still be able to distinguish between an anonymous class/function having been passed to the method and a parse error, the method will continue to output `null` when an anonymous class/function has been passed but will return an empty string when a parse error has been encountered.

Included additional unit tests.