Refactor: add new Util\Sniffs\Comments class
Move comment related utility functions to a separate class.

This deprecates the `Util\Common::$allowedTypes` property and the `Util\Common::suggestType()` function in favour of the same in a new `Util\Sniffs\Comments` class.

This functionality is only used by a select few comment related sniff, so is better placed in a dedicated class.
This also allows for additional comment related functions to be added to this dedicated class.

The deprecated features should be removed in PHPCS 4.0.