Refactor/Comments: adjust $allowedTypes and suggestType() to allow for short form preference
Until now, the `suggestType()` method would always return "long-form" types for param/vars.
However, there are plenty of user-land standards which prefer param/var types to be "short-form".

This commit adjusts the `$allowedTypes` array to have the short-form as keys, while retaining the long-form as values, thus not breaking BC.

It also adds a new `$form` parameter to the `suggestType()` method. The parameter defaults to `short`, as that will be the more common preference nowadays.
To not break BC, in all existing function calls to this method within PHPCS, the parameter is passed as `long`.

Short-form/Long-form only affects the following types:
* `boolean` vs `bool`
* `integer` vs `int`

Other types are not affected.

Includes additional unit tests.