Refactor/Comments: add support for PSR-5 union/intersect types via new suggestTypeString() method
The draft for PSR-5 suggests that union and intersecting types should be supported.
Ref: https://github.com/phpDocumentor/fig-standards/blob/master/proposed/phpdoc.md#details

This commit takes care of that.

The new `suggestTypeString()` method allows for passing a complete type string without pre-processing and receiving a suggested valid type string back.

In combination with the changes made to the `suggestType()` method, the method will:
* Use the same basic type conversions (`real` to `float`, uppercase to lowercase, long to short type etc) as `suggestType`.
* Will recognize single type array `int[]` and multi-type array `(int|float)[]` formats.
* Will, if necessary, unduplicate a type string, i.e `integer|int` will become `int`.
* Will, if necessary, combine multiple PSR-5 style single/multi-type arrays to one multi-type array, i.e. `(int|float)[]|string[]` will become `(int|float|string)[]`.
* Will, if necessary, remove duplicate union/intersect separators.

Includes dedicated unit tests.