Refactor/Comments: add support for nullable types in suggestType() and suggestTypeString() methods
Allows for types to be prefixed with a `?` to indicate nullability.

Note: it will *not* "correct" this to `|null`.
It will, however, now examine the name of the type which was prefixed and suggest corrections if needed.

If, however, both a nullable type _as well as_ `|null` are found, the nullability will be removed from the final suggestion.

Includes unit tests.