Refactor/Comments: implement `suggestTypeString()` in Squiz/FunctionComment [2] / bugfix
Implement the use of the new `Comments::suggestTypeString()` method in the `Squiz.Commenting.FunctionComment` sniff for the checking of `@param` types.

This fixes false negatives for types we are able to correct.

To maintain BC, nullable indicators are not allowed in the types for this sniff.

Includes additional unit tests.