Refactor/Comments: implement findOOStructureComment() in Squiz/ClassComment / bugfix
Implement the use of the new `Comments::findOOStructureComment()` method in the `Squiz.Commenting.ClassComment` sniff.

This should make the finding of the class comment more reliable and consistent and fixes a bug where unrelated trailing end comments would be recognized as if they were class comments.

Includes unit test for the bug fix.