Refactor/Comments: implement findPropertyComment() in Squiz/VariableComment / bugfix
Implement the use of the new `Comments::findPropertyComment()` method in the `Squiz.Commenting.VariableComment` sniff.

This should make the finding of the property comment more reliable and consistent and fixes a bug where unrelated trailing end comments would be recognized as if they were property comments.

Includes unit test for the bug fix.