Refactor/Comments: implement findFunctionComment() in PEAR/Squiz/FunctionComment / bugfix
Implement the use of the new `Comments::findFunctionComment()` method in the `PEAR.Commenting.FunctionComment` sniff and by extension the `Squiz.Commenting.FunctionComment` sniff.

This should make the finding of the function comment more reliable and consistent and fixes a bug where unrelated trailing end comments and comments interlaced into the function declaration would be recognized as if they were function comments.

Includes unit tests for the bug fix.