Refactor/Comments: implement findFunctionComment() in Squiz/FunctionCommentThrowTag / bugfix
Implement the use of the new `Comments::findFunctionComment()` method in the `Squiz.Commenting.FunctionCommentThrowTag` sniff.

This should make the finding of the function comment more reliable and consistent and fixes a bug where unrelated trailing end comments and comments interlaced into the function declaration would be recognized as if they were function comments.

Includes unit test for the bug fix.