Refactor/Comments: implement findFunctionComment() in MySource/AjaxNullComparison / bugfix
Supersedes 2368

Implement the use of the new `Comments::findFunctionComment()` method in the `MySource.Commenting.FunctionComment` sniff.

This should make the finding of the function comment more reliable and consistent and fixes a bug where the sniff could recognize the docblock from a previous function as if it related to the current function.

Includes unit test for the bug fix.

Closes 2368