PSR-12: update the ruleset to reflect current version as under review
As PSR-12 is now in review again, an update of the ruleset to reflect the current version was needed.

* Updated all inline documentation texts with the texts as currently in PSR-12, as it is in review.
* Added any new statements which were added to PSR-12.
* Added the `Generic.WhiteSpace.IncrementDecrementSpacing` sniff to satisfy the `There MUST NOT be any whitespace between the increment/decrement operators and the variable being incremented/decremented.` rule.
* Added the `Squiz.WhiteSpace.CastSpacing` sniff to satisfy the `There MUST NOT be any spaces inside the type casting parentheses.` rule.

Note: The `5.3.1`/`5.3.2` split is my doing as it is otherwise completely unclear what the difference is between the two statements.

While updating the inline documentation, I noticed a few issues in PSR-12 itself and have opened PRs to fix those. If and when those would be accepted, the inline documentation in the ruleset will need to be updated (again).