Generator/Text: bug fix/PHP warning
If an individual line of the code sample in the code comparison would be longer than 47/48 characters, a `Warning: str_repeat(): Second argument has to be greater than or equal to 0` would be thrown.

This fixes the error.

Note: this fix does not introduce code-wrapping as that would probably reduce the readability of the code too much.
Instead, the line will just be longer than other lines.

Idea: should this class respect a `report_width` if set by the user either in their `CodeSniffer.conf` or via the command-line ?