Correct more Error list line numbers
Round 2 of correcting the line numbers for the Error list
Expected errors should now be correct... 
LINE 225 is reporting expected 1 error but found 0 errors. but there is an indent error as the line should be indented more to match the control structure. Seems like line 225 is thinking it belongs to the closure function and not the control structure.