:sparkles: New Generic.WhiteSpace.SpreadOperatorSpacingAfter sniff
This introduces a new sniff which checks the whitespace between the spread operator and the variable/function call it applies to.

This new sniff has two `public` properties:
* `$spacing` which defaults to `0`, i.e. no space.
* `$ignoreNewlines` which defaults to `false`.

Includes unit tests.
Includes fixers.
Includes documentation.

Note: PSR12 only specifies _There MUST NOT be a space between the variadic three dot operator and the argument name_ for **function declarations**.

This sniff, however, checks the spacing for the spread operator in all circumstances in which it can be used, i.e.:
* In function declarations (PHP 5.6).
* In function calls (PHP 5.6).
* In array declarations (PHP 7.4).