Tokenizer/PHP: fix recognition of T_NULLABLE
The `static` keyword is both used as a modifier keyword for functions and variables as well as for late static binding.

The tokenizer did not take late static binding into account when determining whether a `?` is a `T_NULLABLE` token and because of that would make the determination between `T_NULLABLE` and `T_INLINE_THEN` too early.

This has now been fixed.

Includes unit test via the `PSR12.Functions.NullableTypeDeclaration` sniff.