Squiz/DisallowMultipleAssignments: fix ignoring of property declarations
The `Squiz.PHP.DisallowMultipleAssignments` intended to ignore property declarations, but did not allow for typed properties.

I've now moved the "is this a property ?" check up to bow out earlier for all properties.

Includes unit test.

I've also added tests for multi-property declarations. While it is debatable whether or not this sniff should report on these, the existing behaviour was to ignore them. This behaviour has been maintained and is now documented and safeguarded via the test.

Fixes 2787