PHP Tokenizer: improve arrow function backfill consistency
This commit implements the proposal as outlined in 2859#issuecomment-583695917.

The `fn` keyword will now only be tokenized as `T_FN` if it really is an arrow function, in which case it will also have the additional indexes set in the token array.

In all other cases - including when the keyword is used in a function declaration for a named function or method -, it will be tokenized as `T_STRING`.

Includes numerous additional unit tests.

Includes removing the changes made to the `File::getDeclarationName()` function and the `AbstractPatternSniff` in commit 37dda44ed3bf34dbaad6e0dd1e9e10b03d67ed00 as those are no longer necessary.

Fixes #2859

Fixed tests for #2860

Tokenizer arrow functions backfill tests: fix it ;-)

Follow up on 2860. This fixes the tests.

* Adds tests with mixed case `fn` keyword.
* Removes a few redundant tests (weren't testing anything which wasn't covered already).

* Adds a couple of tests which were missing from the data provider.
* Updates the test itself to use a `$testContent` parameter to allow for the `fn` in the tests being non-lowercase.

Added a few more tests for the T_FN backfill (ref #2863, #2860, #2859)