Tokenizer/PHP: support PHP8 union (return) types
`array` keywords used as return types in PHP 8 union types would only be correctly changed to `T_STRING` if they were the first type in the union.

Fixed now.

Includes adding `T_STATIC` to the array of allowed tokens. While previously it wasn't an issue that the token was not includes in the array, it is necessary for the token to be there to support union types.